+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/core/resources/types/TypeNameCollector.java
@@ -79,11 +79,11 @@ public class TypeNameCollector extends TypeNameRequestor {
 	private boolean _requireTypeInProject;
 
 	public TypeNameCollector(IJavaProject project, boolean requireTypeInProject) throws JavaModelException {
-		this(ParsleyProject.getElementClass(project), project, requireTypeInProject, new TreeSet<String>());
+		this(ParsleyProject.getElementClass(project.getProject()), project, requireTypeInProject, new TreeSet<String>());
 	}
 
 	public TypeNameCollector(IJavaProject project, boolean requireTypeInProject, Set<String> typeNames) throws JavaModelException {
-		this(ParsleyProject.getElementClass(project), project, requireTypeInProject, typeNames);
+		this(ParsleyProject.getElementClass(project.getProject()), project, requireTypeInProject, typeNames);
 	}
 	
 	public TypeNameCollector(String superclassTypeName, IJavaProject project, boolean requireTypeInProject) throws JavaModelException {