+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/wodclipse/core/builder/WodBuilder.java
@@ -146,7 +146,7 @@ public class WodBuilder extends AbstractFullAndIncrementalBuilder {
 					if (compilationUnit != null) {
 						IType type = compilationUnit.findPrimaryType();
 						if (type != null) {
-							IType woElementType = type.getJavaProject().findType(ParsleyProject.getElementClass(type.getJavaProject()), progressMonitor);
+							IType woElementType = type.getJavaProject().findType(ParsleyProject.getElementClass(type.getJavaProject().getProject()), progressMonitor);
 							if (woElementType != null) {
 								ITypeHierarchy typeHierarchy = SuperTypeHierarchyCache.getTypeHierarchy(type, progressMonitor);
 								if (typeHierarchy != null && typeHierarchy.contains(woElementType)) {